+++ src/App.js
@@ -1,17 +1,18 @@
+// Dependencies
 import React, { useState } from "react";
 import {BrowserRouter as Router, Routes, Route} from "react-router-dom";
 import './App.css';
-import RelatorioProduzidos from "./Components/RelatorioProduzidos";
-import RelatorioMateriaPrima from "./Components/RelatorioMateriaPrima";
-import Cadastro from './Components/Cadastro';
-import Home from "./Components/Home";
-import Producao from "./Components/producao";
-import Venda from "./Components/Venda";
+
+// Components
 import CadastroMateriaPrima from './Components/CadastroMateriaPrima';
-import RelatorioVenda from "./Components/RelatorioVenda";
+import CadastroProduto from './Components/CadastroProduto';
+import Home from "./Components/Home";
 import Navbar from "./Components/Navbar";
-
-
+import Producao from "./Components/Producao";
+import RelatorioMateriaPrima from "./Components/RelatorioMateriaPrima";
+import RelatorioProduto from "./Components/RelatorioProduto";
+import RelatorioVenda from "./Components/RelatorioVenda";
+import Venda from "./Components/Venda";
 
 
 const App = () => {
@@ -66,11 +67,11 @@ const App = () => {
           <Routes>
             <Route path="/" element={<Home tableMaterial={tableMaterial}/>} />
             <Route path="/home" element={<Home tableMaterial={tableMaterial} />} />
-            <Route path="/cadastro" element={<Cadastro tableMaterial={tableMaterial}/>} />
+            <Route path="/cadastro" element={<CadastroProduto tableMaterial={tableMaterial}/>} />
             <Route path="/cadastroMP" element={<CadastroMateriaPrima onChange={onChange} formInput={formInput} onSubmit={onSubmit}/>} />
             <Route path="/producao" element={<Producao />} />
             <Route path="/venda" element={<Venda />} />
-            <Route path="/relatorioproduzidos" element={<RelatorioProduzidos />} />
+            <Route path="/RelatorioProduto" element={<RelatorioProduto />} />
             <Route path="/relatorioMP" element={<RelatorioMateriaPrima tableMaterial = {tableMaterial}/>} />
             <Route path="/relatorioVenda" element={<RelatorioVenda />} />
           </Routes>
